+++ src/Ext/Rules/Body.h
@@ -370,6 +370,9 @@ public:
 		Valueable<bool> RemoveMindControl_Silent;
 		Valueable<bool> MindControl_Permanent_ReplaceSilent;
 
+		Valueable<DynamicTeamDelayType> TeamDelays_DynamicType;
+		Valueable<Vector3D<int>> TeamDelays_Count[8];
+
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
 			, Storage_TiberiumIndex { -1 }
 			, HarvesterDumpAmount { 0.0f }
@@ -413,6 +416,7 @@ public:
 			, Shield_ConditionYellow { }
 			, Shield_ConditionRed { }
 			, Pips { { 16,17,18 } }
+			, Pips_Shield { { -1,-1,-1 } }
 			, Pips_Shield_Background { }
 			, Pips_Building { { 1,2,4 } }
 			, Pips_Shield_Building { { -1,-1,-1 } }
@@ -674,6 +678,9 @@ public:
 			, DiscardOn_MoveBasedOnDestination { false }
 			, RemoveMindControl_Silent { false }
 			, MindControl_Permanent_ReplaceSilent { false }
+
+			, TeamDelays_DynamicType { DynamicTeamDelayType::StartingPoint }
+			, TeamDelays_Count {}
 		{ }
 
 		virtual ~ExtData() = default;